// src/Types/runners-result.types.ts
export interface RunnersResult {
  id: string;
  timestamp: string;
  currentAsOf: string;
  currentAsOfIncremental: number;
  hisaCriteriaRace: boolean;
  locationId: string;
  raceClaimingPrice: number;
  raceCondition: string;
  raceCountry: string;
  raceCourse: string;
  raceDate: string;
  raceDescription: string;
  raceDistance: string;
  raceDistanceFurlong: string;
  raceDistanceFurlongRound: number;
  raceFootnote: string;
  raceOffTime: string;
  racePostTime: string;
  racePurse: number;
  raceRaceNumber: number;
  raceTrackId: string;
  raceTrackName: string;
  raceType: string;
  starterClaimedPriceUsa: string;
  starterClaimIndicator: string;
  starterClaimingPriceWaived: string;
  starterDnf: boolean;
  starterEarnings: number;
  starterHisaRegulated: boolean;
  starterHorseBreedType: string;
  starterHorseColor: string;
  starterHorseDamName: string;
  starterHorseFoaled: string;
  starterHorseHisaId: string;
  starterHorseMicrochips: string;
  starterHorseName: string;
  starterHorseReferenceNumber: number;
  starterHorseRegistry: string;
  starterHorseSex: string;
  starterHorseSireName: string;
  starterHorseTattoo: string;
  starterJockeyFirstName: string;
  starterJockeyHisaId: string;
  starterJockeyLastName: string;
  starterJockeyReferenceNumber: number;
  starterJockeyType: string;
  starterLengthBehindAtFinish: number;
  starterOdds: number;
  starterOfficialPosition: number;
  starterOwnerFirstName: string;
  starterOwnerHisaId: string;
  starterOwnerLastName: string;
  starterOwnerReferenceNumber: number;
  starterOwnerType: string;
  starterPositionAtPointOfCall1: number;
  starterPositionAtPointOfCall2: number;
  starterPositionAtPointOfCall3: number;
  starterPositionAtPointOfCall4: number;
  starterPositionAtPointOfCall5: number;
  starterPostPosition: number;
  starterProgramNumber: string;
  starterRecordType: string;
  starterScratched: boolean;
  starterShakes: number;
  starterTrainerFirstName: string;
  starterTrainerHisaId: string;
  starterTrainerLastName: string;
  starterTrainerReferenceNumber: number;
  starterTrainerType: string;
  starterVoidIndicator: string;
  starterVoidReason: string;
}

export interface RunnersResultResponse {
  items: RunnersResult[];
  nextToken?: string;
  count: number;
  total?: number;
  totalCount?: number;
  currentPage?: number;
  pageSize?: number;
  totalPages?: number;
  hasNextPage?: boolean;
  hasPreviousPage?: boolean;
}